Add withdrawal and deposit handlers with appropriate testing.

+++ internal/migrations/20220730173930_create_account.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS accounts
 (
 	id 				UUID 		PRIMARY KEY DEFAULT gen_random_uuid(),	
 	name         	VARCHAR(50) NOT NULL,
-	balance			INTEGER 	NOT NULL,
+	balance			BIGINT  NOT NULL,
 	created_at 		TIMESTAMP 	NOT NULL,
 	last_modified   TIMESTAMP   NOT NULL
 );
